+++ tests/tests_highscores.py
@@ -7,35 +7,34 @@ from tibiapy import HighscoresCategory, InvalidContent, \
 from tibiapy.models import Highscores, HighscoresEntry, LoyaltyHighscoresEntry
 from tibiapy.parsers.highscores import HighscoresParser
 
-FILE_HIGHSCORES_FULL = "highscores/tibiacom_full.txt"
-FILE_HIGHSCORES_EXPERIENCE = "highscores/tibiacom_experience.txt"
-FILE_HIGHSCORES_LOYALTY = "highscores/tibiacom_loyalty.txt"
-FILE_HIGHSCORES_BATTLEYE_PVP_FILTER = "highscores/tibiacom_battleye_pvp_filters.txt"
-FILE_HIGHSCORES_EMPTY = "highscores/tibiacom_empty.txt"
-FILE_HIGHSCORES_NO_RESULTS = "highscores/tibiacom_no_results.txt"
+FILE_HIGHSCORES_FULL = "highscores/highscores.txt"
+FILE_HIGHSCORES_EXPERIENCE = "highscores/highscoresExperience.txt"
+FILE_HIGHSCORES_LOYALTY = "highscores/highscoresLoyalty.txt"
+FILE_HIGHSCORES_BATTLEYE_PVP_FILTER = "highscores/highscoresBattleEyePvpFilters.txt"
+FILE_HIGHSCORES_EMPTY = "highscores/highscoresEmpty.txt"
+FILE_HIGHSCORES_NO_RESULTS = "highscores/highscoresNoResults.txt"
 
 
 class TestHighscores(unittest.TestCase, TestCommons):
     # region Tibia.com Tests
-    def test_highscores_from_content(self):
+    def test_highscores_parser_from_content(self):
         """Testing parsing Highscores"""
         content = self.load_resource(FILE_HIGHSCORES_FULL)
         highscores = HighscoresParser.from_content(content)
 
-        self.assertEqual("Estela", highscores.world)
+        self.assertEqual("Gladera", highscores.world)
         self.assertEqual(HighscoresProfession.KNIGHTS, highscores.vocation)
         self.assertEqual(HighscoresCategory.MAGIC_LEVEL, highscores.category)
         self.assertEqual(HighscoresBattlEyeType.ANY_WORLD, highscores.battleye_filter)
-        self.assertEqual(1983, highscores.results_count)
-        self.assertEqual(38, highscores.from_rank)
-        self.assertEqual(38, highscores.to_rank)
+        self.assertEqual(1688, highscores.results_count)
+        self.assertEqual(16, highscores.from_rank)
+        self.assertEqual(168, highscores.to_rank)
         self.assertEqual(4, highscores.current_page)
-        self.assertEqual(40, highscores.total_pages)
+        self.assertEqual(34, highscores.total_pages)
         self.assertIsNotNone(highscores.get_page_url(1))
         self.assertIsNotNone(highscores.url)
         self.assertIsNotNone(highscores.next_page_url)
         self.assertIsNotNone(highscores.previous_page_url)
-        self.assertEqual(datetime.timedelta(minutes=6), highscores.last_updated)
 
         for entry in highscores.entries:
             self.assertIsInstance(entry, HighscoresEntry)
@@ -44,9 +43,9 @@ class TestHighscores(unittest.TestCase, TestCommons):
             self.assertIsInstance(entry.rank, int)
             self.assertIsInstance(entry.value, int)
             self.assertIsInstance(entry.level, int)
-            self.assertEqual("Estela", entry.world)
+            self.assertEqual("Gladera", entry.world)
 
-    def test_highscores_from_content_highscores(self):
+    def test_highscores_parser_from_content_experience(self):
         """Testing parsing experience highscores"""
         content = self.load_resource(FILE_HIGHSCORES_EXPERIENCE)
         highscores = HighscoresParser.from_content(content)
@@ -66,7 +65,7 @@ class TestHighscores(unittest.TestCase, TestCommons):
             self.assertIsInstance(entry.level, int)
             self.assertEqual(entry.world, "Gladera")
 
-    def test_highscores_from_content_loyalty(self):
+    def test_highscores_parser_from_content_loyalty(self):
         """Testing parsing experience loyalty"""
         content = self.load_resource(FILE_HIGHSCORES_LOYALTY)
         highscores = HighscoresParser.from_content(content)
@@ -74,8 +73,8 @@ class TestHighscores(unittest.TestCase, TestCommons):
         self.assertEqual("Calmera", highscores.world)
         self.assertEqual(HighscoresProfession.PALADINS, highscores.vocation)
         self.assertEqual(HighscoresCategory.LOYALTY_POINTS, highscores.category)
-        self.assertEqual(1007, highscores.results_count)
-        self.assertEqual(21, highscores.total_pages)
+        self.assertEqual(1000, highscores.results_count)
+        self.assertEqual(20, highscores.total_pages)
 
         for entry in highscores.entries:
             self.assertIsInstance(entry, LoyaltyHighscoresEntry)
@@ -86,7 +85,7 @@ class TestHighscores(unittest.TestCase, TestCommons):
             self.assertIsInstance(entry.level, int)
             self.assertIsInstance(entry.title, str)
 
-    def test_highscores_from_content_battleye_and_pvp_filters(self):
+    def test_highscores_parser_from_content_battleye_and_pvp_filters(self):
         """Testing parsing Highscores"""
         content = self.load_resource(FILE_HIGHSCORES_BATTLEYE_PVP_FILTER)
         highscores = HighscoresParser.from_content(content)
@@ -100,9 +99,8 @@ class TestHighscores(unittest.TestCase, TestCommons):
         self.assertEqual(50, highscores.to_rank)
         self.assertEqual(1, highscores.current_page)
         self.assertEqual(20, highscores.total_pages)
-        self.assertEqual(3, len(highscores.pvp_types_filter))
+        self.assertEqual(4, len(highscores.pvp_types_filter))
         self.assertIsNotNone(highscores.url)
-        self.assertEqual(datetime.timedelta(minutes=27), highscores.last_updated)
 
         for entry in highscores.entries:
             self.assertIsInstance(entry, HighscoresEntry)
@@ -112,26 +110,25 @@ class TestHighscores(unittest.TestCase, TestCommons):
             self.assertIsInstance(entry.value, int)
             self.assertIsInstance(entry.level, int)
 
-    def test_highscores_from_content_empty(self):
+    def test_highscores_parser_from_content_empty(self):
         """Testing parsing highscores when empty (world doesn't exist)"""
         content = self.load_resource(FILE_HIGHSCORES_EMPTY)
         highscores = HighscoresParser.from_content(content)
 
         self.assertIsNone(highscores)
 
-    def _test_highscores_from_content_no_results(self):
+    def test_highscores_parser_from_content_no_results(self):
         """Testing parsing highscores with no results (first day of a new world)"""
         content = self.load_resource(FILE_HIGHSCORES_NO_RESULTS)
         highscores = HighscoresParser.from_content(content)
 
         self.assertIsInstance(highscores, Highscores)
-        self.assertEqual(highscores.world, "Unica")
-        self.assertEqual(highscores.category, HighscoresCategory.EXPERIENCE)
-        self.assertEqual(highscores.vocation, HighscoresProfession.ALL)
-        self.assertEqual(highscores.total_pages, 0)
-        self.assertEqual(len(highscores.entries), 0)
+        self.assertEqual("Gravitera", highscores.world)
+        self.assertEqual(1, highscores.total_pages)
+        self.assertEqual(0, len(highscores.entries))
+        self.assertEqual(0, highscores.results_count)
 
-    def test_highscores_from_content_unrelated_section(self):
+    def test_highscores_parser_from_content_unrelated_section(self):
         """Testing parsing an unrelated section"""
         content = self.load_resource(self.FILE_UNRELATED_SECTION)
         with self.assertRaises(InvalidContent):